requirements.txt
discord.py==2.5.2
requests==2.31.0
beautifulsoup4==4.13.3
gspread==6.2.1
oauth2client==4.1.3
schedule==1.2.2
lxml==5.1.0
selenium==4.15.0
webdriver-manager==4.0.1
python-dotenv==1.0.1
